requirements.txt
# Core dependencies for Telegram File-to-Link Bot
pyrogram==2.0.106
aiohttp==3.9.1
aiofiles==23.2.1

# Cryptography and security
tgcrypto==1.2.5
cryptography==41.0.8

# Environment and configuration
python-dotenv==1.0.0

# Async utilities
asyncio-throttle==1.0.2

# Logging and monitoring
colorlog==6.8.0

# HTTP client for additional requests
httpx==0.25.2

# Memory optimization
psutil==5.9.6

# Date and time utilities
python-dateutil==2.8.2

# JSON handling improvements
orjson==3.9.10

# Additional utilities
humanize==4.9.0
